Resolve dropped CVPixelBuffers in MovieOutput
https://developer.apple.com/documentation/avfoundation/avassetwriterinputpixelbufferadaptor/1388102-append
Documentation: "Do not modify a CVPixelBuffer or its contents after you have passed it to this method."

It must then be a requirement that for each frame we create a new pixel buffer with CVPixelBufferPoolCreatePixelBuffer(). Since we are not doing this, occasionally old pixel buffers are being reused. This is causing dropped frames - in the form of duplicates. These would mostly occur towards the beginning of videos and some would be interspersed throughout the rest of the video. I also noticed this problem impacted certain devices more so than others, usually older devices more than newer devices. With this resolved video encoding is now a 1-1 mapping of frames and 60fps video is supported without issue.

https://github.com/BradLarson/GPUImage/issues/1501