Fix issue where glReadPixels did not account for pixel buffer padding in simulator.
Fix issue where glReadPixels did not account for pixel buffer padding in simulator. This fixes a distortion issue in the simulator that would occur when the video width was not a multiple of 16.
Removed colorSwizzlingShader since we can just read the pixels in the format of our choosing with glReadPixels.